+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## actions development version
 
-- `mkdocs-mike`: new option `strict-semver` to toggle whether to enforce strict semantic versioning or allow a relaxed format with only the major and minor components. (#)
+- `mkdocs-mike`: new option `strict-semver` to toggle whether to enforce strict semantic versioning or allow a relaxed format with only the major and minor components. (#72, @kelly-sovacool)
 
 ## actions 0.2.6